Move base.py to common folder
