+++ src/llm/common/__init__.py
@@ -1,0 +1,12 @@
+"""
+Common LLM infrastructure components.
+
+This module contains shared components used across different LLM providers,
+including the base client class and common utilities.
+"""
+
+from src.llm.common.base import BaseLLMClient
+
+__all__ = [
+    "BaseLLMClient",
+]
